Restructure for pytest
